Separate blinky demo from main code
