+++ FreeRTOS/Demo/RISC-V-Qemu-virt_GCC/main.c
@@ -28,144 +28,30 @@
 /* FreeRTOS kernel includes. */
 #include <FreeRTOS.h>
 #include <task.h>
-#include <queue.h>
 
-#include <stdio.h>
+/* Run a simple demo just prints 'Blink' */
+#define DEMO_BLINKY	1
 
-#include "riscv-virt.h"
-#include "ns16550.h"
-
-/* Priorities used by the tasks. */
-#define mainQUEUE_RECEIVE_TASK_PRIORITY		( tskIDLE_PRIORITY + 2 )
-#define	mainQUEUE_SEND_TASK_PRIORITY		( tskIDLE_PRIORITY + 1 )
-
-/* The rate at which data is sent to the queue.  The 200ms value is converted
-to ticks using the pdMS_TO_TICKS() macro. */
-#define mainQUEUE_SEND_FREQUENCY_MS			pdMS_TO_TICKS( 1000 )
-
-/* The maximum number items the queue can hold.  The priority of the receiving
-task is above the priority of the sending task, so the receiving task will
-preempt the sending task and remove the queue items each time the sending task
-writes to the queue.  Therefore the queue will never have more than one item in
-it at any time, and even with a queue length of 1, the sending task will never
-find the queue full. */
-#define mainQUEUE_LENGTH					( 1 )
-
-/*
- * Prototypes for the standard FreeRTOS callback/hook functions implemented
- * within this file.  See https://www.freertos.org/a00016.html
- */
 void vApplicationMallocFailedHook( void );
 void vApplicationIdleHook( void );
 void vApplicationStackOverflowHook( TaskHandle_t pxTask, char *pcTaskName );
 void vApplicationTickHook( void );
 
-/*-----------------------------------------------------------*/
-
-/* The queue used by both tasks. */
-static QueueHandle_t xQueue = NULL;
-
-/*-----------------------------------------------------------*/
-
-static void prvQueueSendTask( void *pvParameters )
-{
-TickType_t xNextWakeTime;
-const unsigned long ulValueToSend = 100UL;
-const char * const pcMessage1 = "Transfer1";
-const char * const pcMessage2 = "Transfer2";
-int f = 1;
-
-	/* Remove compiler warning about unused parameter. */
-	( void ) pvParameters;
-
-	/* Initialise xNextWakeTime - this only needs to be done once. */
-	xNextWakeTime = xTaskGetTickCount();
-
-	for( ;; )
-	{
-		char buf[40];
-
-		sprintf( buf, "%d: %s: %s", xGetCoreID(),
-				pcTaskGetName( xTaskGetCurrentTaskHandle() ),
-				( f ) ? pcMessage1 : pcMessage2 );
-		vSendString( buf );
-		f = !f;
-
-		/* Place this task in the blocked state until it is time to run again. */
-		vTaskDelayUntil( &xNextWakeTime, mainQUEUE_SEND_FREQUENCY_MS );
-
-		/* Send to the queue - causing the queue receive task to unblock and
-		toggle the LED.  0 is used as the block time so the sending operation
-		will not block - it shouldn't need to block as the queue should always
-		be empty at this point in the code. */
-		xQueueSend( xQueue, &ulValueToSend, 0U );
-	}
-}
-
-/*-----------------------------------------------------------*/
-
-static void prvQueueReceiveTask( void *pvParameters )
-{
-unsigned long ulReceivedValue;
-const unsigned long ulExpectedValue = 100UL;
-const char * const pcMessage1 = "Blink1";
-const char * const pcMessage2 = "Blink2";
-const char * const pcFailMessage = "Unexpected value received\r\n";
-int f = 1;
-
-	/* Remove compiler warning about unused parameter. */
-	( void ) pvParameters;
-
-	for( ;; )
-	{
-		char buf[40];
-
-		/* Wait until something arrives in the queue - this task will block
-		indefinitely provided INCLUDE_vTaskSuspend is set to 1 in
-		FreeRTOSConfig.h. */
-		xQueueReceive( xQueue, &ulReceivedValue, portMAX_DELAY );
-
-		/*  To get here something must have been received from the queue, but
-		is it the expected value?  If it is, toggle the LED. */
-		if( ulReceivedValue == ulExpectedValue )
-		{
-			sprintf( buf, "%d: %s: %s", xGetCoreID(),
-					pcTaskGetName( xTaskGetCurrentTaskHandle() ),
-					( f ) ? pcMessage1 : pcMessage2 );
-			vSendString( buf );
-			f = !f;
-
-			ulReceivedValue = 0U;
-		}
-		else
-		{
-			vSendString( pcFailMessage );
-		}
-	}
-}
+int main_blinky( void );
 
 /*-----------------------------------------------------------*/
 
 int main( void )
 {
-	vSendString( "Hello FreeRTOS!" );
+	int ret;
 
-	/* Create the queue. */
-	xQueue = xQueueCreate( mainQUEUE_LENGTH, sizeof( uint32_t ) );
+#if defined(DEMO_BLINKY)
+	ret = main_blinky();
+#else
+#error "Please add or select demo."
+#endif
 
-	if( xQueue != NULL )
-	{
-		/* Start the two tasks as described in the comments at the top of this
-		file. */
-		xTaskCreate( prvQueueReceiveTask, "Rx", configMINIMAL_STACK_SIZE * 2U, NULL,
-					mainQUEUE_RECEIVE_TASK_PRIORITY, NULL );
-		xTaskCreate( prvQueueSendTask, "Tx", configMINIMAL_STACK_SIZE * 2U, NULL,
-					mainQUEUE_SEND_TASK_PRIORITY, NULL );
-	}
-
-	vTaskStartScheduler();
-
-	return 0;
+	return ret;
 }
 
 /*-----------------------------------------------------------*/
